% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_data_extraction.R
\name{fill_coastal_cells}
\alias{fill_coastal_cells}
\title{Fill missing cells}
\usage{
fill_coastal_cells(r, w = 3, max_iter = 50, mask = depth)
}
\arguments{
\item{r}{A raster layer.}

\item{w}{Neighborhood area, must be odd.}

\item{max_iter}{Maximum number to times to iterate until cells are filled.}

\item{mask}{Raster layer with land cells NA.}
}
\value{
Raster layer of dim(r), with all land cells NA and all water cells containing data.
}
\description{
Iteratively applies 3x3 nearest neighbor means to NA cells until all water cells
have data.  Typically used to fill areas near the coastlilne with grids don't align perfectly.
}
\examples{
\dontrun{
fill_coastal_cells(r, mask=depth)
}
}
