% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn.disp_testval_tags}
\alias{fn.disp_testval_tags}
\title{Make dispersal parameter taglines for sensitivity runs.}
\usage{
fn.disp_testval_tags(disp.base, pct.change = 0.5)
}
\arguments{
\item{disp.base}{A dataframe with baseline dispersal parameters, exported from EwE and read and
renamed in the setup file.}

\item{pct.change}{The percent change to use,
applied as pct.change x base and (1+pct.change) x base. Default is 0.5 (+/-50\%).}
}
\value{
A character vector containing the parameter taglines for the command file.
}
\description{
Create parameter lines for the command file to test sensitivity of dispersal rates.
}
\examples{
# example code:
\dontrun{tags.vul <- fn.vul_testval_tags(predprey=data.frame(pred=1:5,prey=6:10, basevul=2), maxvul=1000, minvul=1.01, is.maxvul.mult=T)}
}
