% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_data_extraction.R
\name{fn.download_cefi_full_nc}
\alias{fn.download_cefi_full_nc}
\title{Download full CEFI netcdf files}
\usage{
fn.download_cefi_full_nc(
  vars.cefi = vars.cefi,
  dir.cefi = dir.cefi,
  reg = "northwest_atlantic"
)
}
\arguments{
\item{vars.cefi}{A dataframe returned by fn.get_cefi_vars(), likely subset to fewer variables.
Must contain the variables "cefi_filename" and "cefi_rel_path".}

\item{dir.cefi}{Parent directory for CEFI data, where the netcdf files will be saved.}

\item{reg}{CEFI region, either northwest_atlantic or northeast_pacific}
}
\description{
Downloads the full netcdf for MOM6-COBALT variables from
https://downloads.psl.noaa.gov/Projects/CEFI/regional_mom6.  Uses the httr::GET function.
}
\examples{
# example code:
\dontrun{vars.cefi <- fn.get_cefi_vars(dir.cefi=dir.cefi, file.cefivars='CEFI_vars.csv', region='nwa')}
}
