% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{fn.ecosim_predB_ts2array}
\alias{fn.ecosim_predB_ts2array}
\title{Read Ecosim predicted biomass timeseries output.}
\usage{
fn.ecosim_predB_ts2array(dir.out = dir.pred, timestep = "annual")
}
\arguments{
\item{dir.out}{Output directory.  All .csv biomass files created with Ecospace naming conventions
in this directory will be read.  Can be a vector of directories.}

\item{timestep}{Read 'annual' or 'monthly' data.}
}
\value{
An array of biomass output with dimensions (nyrs,ngrps,nregions,nruns)
#examples
#example code:
predB <- fn.ecospace_predB_ts2array (dir.out="./output/", timestep='annual',n.reg=0)
}
\description{
Reads biomass timeseries output csv into an array.
}
