% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{fn.ecospace_ascii2stack}
\alias{fn.ecospace_ascii2stack}
\title{Read Ecospace output ascii files into a raster stack.}
\usage{
fn.ecospace_ascii2stack(
  dir.out = dir.pred,
  do.vars = c("biomass"),
  do.grps = 1:12,
  do.fleets = 1:6,
  do.months = 1,
  do.years = 2010:2019,
  annualmaps = TRUE
)
}
\arguments{
\item{dir.out}{The run output folder containing.  It should contain a folder named 'asc', which
holds all the output maps.}

\item{do.vars}{The model output variables to read. Must match the naming convention of the output
files, all  lowercase.  Current options are 'biomass','catch', and 'effort'.}

\item{do.grps}{The model groups for which to read data.}

\item{do.fleets}{The fleet numbers for which to read effort data.}

\item{do.months}{The months to read output data, only used if monthly maps are output.}

\item{do.years}{Years to read output data, will combine with do.months}

\item{annualmaps}{Import annual (TRUE, default) or monthly (FALSE) maps from Ecospace.}
}
\value{
A SpatRaster stack from the terra package.
}
\description{
This function reads a subset of asc grids produced by an Ecospace run into a raster
stack.  Requires the terra package for faster processing.
}
\examples{
# example code:
\dontrun{predB.maps <- fn.ecospace_ascii2stack(dir.out=dir.pred, do.vars=c('biomass'), do.grps=c(3,5,6,8,10,12), do.fleets=1:6, do.months=1, do.years=2010:2019, annualmaps=TRUE)}
}
