% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{fn.ecospace_plot_ts}
\alias{fn.ecospace_plot_ts}
\title{Plot Ecospace predicted timeseries.}
\usage{
fn.ecospace_plot_ts(
  predB = predB,
  predC = predC,
  timestep = "annual",
  obs.ts = obs.ts,
  scale2run = 1,
  pltB.dims = c(3, 3),
  pltC.dims = c(1, 1),
  scaleCatch = FALSE,
  plt.cols = 1:dim(predB)[3],
  plt.obs = TRUE,
  dir.plts = dir.pred,
  plot2pdf = TRUE,
  run.label = Sys.Date()
)
}
\arguments{
\item{predB}{An array of biomass predictions, as produced by fn.ecospace_ts2array()}

\item{predC}{An array of catch predictions, as produced by fn.ecospace_ts2array()}

\item{timestep}{Read 'annual' or 'monthly' data.}

\item{obs.ts}{A list object containing reference timeseries information, as that created by
fn.read_ecosim_timeseries().}

\item{scale2run}{If multiple models, which to scale the observed values to.}

\item{pltB.dims}{Numeric vector of length 2 specifying the number of rows and columns for multipanel plotting.}

\item{pltC.dims}{Numeric vector of length 2 specifying the number of rows and columns for multipanel plotting.}

\item{scaleCatch}{Logical. TRUE will scale the predictions so the mean is equal to that of the observed.  Default is FALSE.}

\item{plt.cols}{Vector of line colors.}

\item{plt.obs}{Logical. Should observed data in obs.ts be plotted?}

\item{dir.plts}{Location to save pdf plots. Defaults to dir.pred.}

\item{plot2pdf}{Logical.  TRUE (default) will save files to the specified directory.}

\item{run.label}{Label appended to the output pdf file name (and used to tag the
run). Defaults to the current date.}
}
\value{
Generate a figure in the active plotting device or saves as pdf file.
}
\description{
Makes a multipanel plot of 1 or more ecopace runs with observed data where available.
}
\examples{
# example code:
\dontrun{result <- fn.runEwE.parallel(runlist=myrunlist, obj.fxn=1, cl.export=list('myrunlist','obs.ts'))}
}
