% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{fn.ecospace_predC_ts2array}
\alias{fn.ecospace_predC_ts2array}
\title{Read Ecospace predicted catch timeseries output.}
\usage{
fn.ecospace_predC_ts2array(dir.out = dir.pred, timestep = "annual", n.reg = 0)
}
\arguments{
\item{dir.out}{Output directory.  All .csv catch files created with Ecospace naming conventions
in this directory will be read.}

\item{timestep}{Read 'annual' or 'monthly' data.}

\item{n.reg}{If output for multiple regions, how many?  Default to 0.}
}
\value{
An array of catch output with dimensions (nyrs,ngrps,nregions,nruns)
#examples
#example code:
predB <- fn.ecospace_predB_ts2array (dir.out="./output/", timestep='annual',n.reg=0)
}
\description{
Reads catch timeseries output csv, sums over fleets for each species, and puts into
an array.
}
