% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Ecospace_GA_calibration.R
\name{fn.GApop}
\alias{fn.GApop}
\title{Generate an initial GA population matrix}
\usage{
fn.GApop(run_config = myconfig, pardist = "uniform", vuldist = "uniform")
}
\arguments{
\item{run_config}{List containing GA settings; must include \code{popSize}.}

\item{pardist}{Character string; distribution type, either"uniform" (uniform between bounds) or
"normal" (normal/gamma hybrid). Default "uniform"}

\item{vuldist}{Character string; distribution type for vulnerability parameters, either"uniform" (uniform between bounds) or
"gamma" withc shape=2 and rate=shape/par. Default "uniform"}
}
\value{
A numeric matrix of dimension run_config$popSize x n_pars.
}
\description{
Creates the initial population for a genetic algorithm using either
uniform draws within parameter bounds or (optionally) normal/gamma
draws centered on existing parameter estimates. Uses several global
objects created elsewhere.
}
