% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_data_extraction.R
\name{fn.get_cefi_vars}
\alias{fn.get_cefi_vars}
\title{List variables and get URLs for CEFI data.}
\usage{
fn.get_cefi_vars(
  dir.cefi = dir.cefi,
  file.cefivars = "CEFI_vars.csv",
  region = "nwa"
)
}
\arguments{
\item{dir.cefi}{Parent directory for CEFI data.}

\item{file.cefivars}{Name of csv file to be created (or read) containing cefi variables. Default
is CEFI_vars.csv.}

\item{region}{CEFI region nwa=northwest Atlantic, nep=northeast Pacific}
}
\description{
This function will get information on CEFI data and return a dataframe that can be
subset for pulling data.
}
\examples{
# example code:
\dontrun{vars.cefi <- fn.get_cefi_vars(dir.cefi=dir.cefi, file.cefivars='CEFI_vars.csv', region='nwa')}
}
