% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Spatial_likelihood.R
\name{fn.gfisher_default_crosswalk}
\alias{fn.gfisher_default_crosswalk}
\title{Default GFISHER mod-code -> WFS MICE pool-code crosswalk.}
\usage{
fn.gfisher_default_crosswalk()
}
\value{
data.frame with columns \code{mod, pc, species}.
}
\description{
gag ages 0-5+ = mod 20-25 = pc 4-9; red grouper ages 0-5+ =
mod 26-31 = pc 10-15. Crosswalk is \code{pc = mod - 16}.
}
