% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn.make_cmd_files}
\alias{fn.make_cmd_files}
\title{Make command files.}
\usage{
fn.make_cmd_files(runlist = runlist, nyrs = nyrs)
}
\arguments{
\item{runlist}{A dataframe where each row is a separate model run.  It must include, at a
minimum, the 3 variables 'dir.out', which provides the output directory for each run;
'cmd_file' which is the full file path of the command file to be saved in the folder; and 'tag',
which is the tagline to be appended to the command file.}

\item{nyrs}{A single number representing the number of years to run ecospace.}
}
\value{
Create command .txt files and saves them in their respecitive run folders.
}
\description{
This function adds taglines to the base command file and saves them in run folders.
}
\examples{
# example code:
\dontrun{fn.make_cmd_files(runlist)}
}
