% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_data_extraction.R
\name{fn.make_monthly_climatology_maps}
\alias{fn.make_monthly_climatology_maps}
\title{Make monthly climatology maps.}
\usage{
fn.make_monthly_climatology_maps(dir.stdriver = dir.stdriver)
}
\arguments{
\item{dir.stdriver}{Directory for ST driver variable.}
}
\value{
Saves 12 ascii files to the variable folder with suffix '9999mm', where mm is numeric
month.
}
\description{
This stacks all the ascii in stdriver directory and compute long term mean for each
month.
}
\examples{
\dontrun{
# example code:
for(i in 1:length(dirs.stvars)){
fn.make_monthly_climatology_maps(dir.stdriver=dirs.stvars[i])
}}
}
