% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_data_extraction.R
\name{fn.make_static_maps}
\alias{fn.make_static_maps}
\title{Make static maps from monthly ST files.}
\usage{
fn.make_static_maps(dir.stdriver = dir.stdriver)
}
\arguments{
\item{dir.stdriver}{Directory for ST driver variable.  A 'static' folder will be created one
level up if it doesn't exist.}
}
\value{
Saves an ascii file to the static folder.
}
\description{
This stacks all the ascii in stdriver directory and compute long term mean and
year-1 mean layers, and saves them as ascii files in the 'static' folder.
}
\examples{
# example code:
\dontrun{fn.make_static_maps(dir.stdriver=file.path(dir.stdriver,"tos_cefi"))}
}
