% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn.mediation_testval_tags}
\alias{fn.mediation_testval_tags}
\title{Make mediation shape parameter taglines for sensitivity runs.}
\usage{
fn.mediation_testval_tags(meds.base, do.xbase = TRUE)
}
\arguments{
\item{meds.base}{A dataframe with baseline mediation parameters, this has to be created manually.}

\item{do.xbase}{Logical.  Should the x_base parameter also be tested? (default=TRUE).}
}
\value{
A character vector containing the parameter taglines for the command file.
}
\description{
Create parameter lines for the command file to test sensitivity to mediation shapes.
}
\examples{
# example code:
\dontrun{tags.vul <- fn.vul_testval_tags(predprey=data.frame(pred=1:5,prey=6:10, basevul=2), maxvul=1000, minvul=1.01, is.maxvul.mult=T)}
}
