% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecospace_objective_functions.R
\name{fn.objfxn1}
\alias{fn.objfxn1}
\title{Ecospace objective function, timeseries only}
\usage{
fn.objfxn1(dir.pred, obs.ts = obs.ts, run.idx = 1, fit.abs.catch = TRUE)
}
\arguments{
\item{dir.pred}{Full directory path to model output folder containing .csv files.}

\item{obs.ts}{A list object containing reference timeseries information, as that created by
fn.read_ecosim_timeseries().}

\item{run.idx}{If dir.pred has multiple output folders, use this to select which to calculate.}

\item{fit.abs.catch}{If TRUE (default), do not rescale predicions.  If FALSE, predictions were
divided by q, where q=mean(pred)/mean(obs)}
}
\value{
A vector containing the total and functional group specific likelihood.
}
\description{
Calculates the composite likelihood.
}
