% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ecospace_objective_functions.R
\name{fn.objfxn2}
\alias{fn.objfxn2}
\title{Ecospace objective function, timeseries plut spatial reference data}
\usage{
fn.objfxn2(
  dir.pred,
  obs.ts = obs.ts,
  obs.maps = obs.maps,
  obs.maps.meta = obs.maps.meta,
  autoweight.LL = T
)
}
\arguments{
\item{dir.pred}{Full directory path to model output folder containing .csv files.}

\item{obs.ts}{A list object containing reference timeseries information, as that created by
fn.read_ecosim_timeseries().}

\item{obs.maps}{A raster stack???}

\item{obs.maps.meta}{A dataframe??? Likely nrow=nlyrs(obs.maps) If TRUE (default), do not rescale
predicions.  If FALSE, predictions were divided by q, where q=mean(pred)/mean(obs)}

\item{autoweight.LL}{Logical value.  If TRUE (default), the timeseries and maps likelihoods are
each scaled to their own means LL.w=LL*1/mean(LL)  This puts them on equal footing, but needs more thought.}
}
\value{
A vector containing the total and functional group specific likelihood.
}
\description{
Calculates the composite likelihood, including spatial reference datasets.
}
\keyword{internal}
