% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_data_extraction.R
\name{fn.pull_cefi}
\alias{fn.pull_cefi}
\title{Pull CEFI data}
\usage{
fn.pull_cefi(
  vars.cefi = vars.cefi,
  dir.cefi = dir.cefi,
  reg = "northwest_atlantic",
  depth = depth.5min,
  delete.full.nc = FALSE
)
}
\arguments{
\item{vars.cefi}{A dataframe returned by fn.get_cefi_vars(), likely subset to fewer variables.
Must contain the variables "cefi_filename" and "cefi_rel_path".}

\item{dir.cefi}{Parent directory for CEFI data, where the netcdf files will be saved.}

\item{reg}{CEFI region, either northwest_atlantic or northeast_pacific}

\item{depth}{Raster grid of depth, matching the basemap of Ecospace.  Saved ascii files will have
the same dimensions.}

\item{delete.full.nc}{Logical. TRUE will delete the full nc files if they exist.  Default is False.}
}
\description{
Downloads the full netcdf for MOM6-COBALT variables from
https://downloads.psl.noaa.gov/Projects/CEFI/regional_mom6 using the httr::GET function.  It
applies the spatial subset based on the extent of depth and writes a new netcdf file.
}
\examples{
# example code:
\dontrun{fn.pull_cefi(vars.cefi=vars.cefi, dir.cefi=dir.cefi, reg='northwest_atlantic', depth=depth.5min, delete.full.nc=TRUE)}
}
