% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ST_data_extraction.R
\name{fn.pull_glorys}
\alias{fn.pull_glorys}
\title{Pull GLORYS}
\usage{
fn.pull_glorys(
  dir.out = dir.glorys,
  bbox = bbox,
  startdate = startdate,
  enddate = enddate,
  datid.phy = "cmems_mod_glo_phy-all_my_0.25deg_P1M-m",
  datid.bgc = "cmems_mod_glo_bgc_my_0.25deg_P1M-m",
  vars.phy = c("thetao", "so", "bottomT"),
  vars.bgc = c("chl", "nppv", "o2", "phyc"),
  path.copernicusmarine = "C:/~/copernicusmarine.exe",
  getbathygrid = FALSE
)
}
\arguments{
\item{dir.out}{Output directory to save netcdf files.}

\item{bbox}{Numeric vector containing spatial domain (W,E,S,N).}

\item{startdate}{First day to pull data, YYYY-MM-DD.}

\item{enddate}{Last day to pull data, YYYY-MM-DD.}

\item{datid.phy}{Dataset id for physical data, visit website.}

\item{datid.bgc}{Dataset id for biogeochemical data, visit website.}

\item{vars.phy}{Character vector containing physical variable names to download.  Check website
for correct names.}

\item{vars.bgc}{Character vector containing biogeochemical variable names to download.  Check
website for correct names.}

\item{path.copernicusmarine}{File path location of copernicusmarine.exe file.}

\item{getbathygrid}{Logical.  Should bathymetric grids be downloaded?}
}
\value{
One or more netcdf files containing the data.
}
\description{
Pull monthly 1/4 degree physical and biogeochemical hindcast products from the
Copernicus marine data store.  Requires the Copernicus marine toolbox CLI application.
Check website https://data.marine.copernicus.eu/products for dataset id and variable names.
}
\examples{
# example code:
\dontrun{fn.pull_glorys(dir.out=dir.glorys, bbox=bbox, startdate = startdate, enddate = enddate, vars.phy = c('thetao','so','bottomT'), vars.bgc = c('chl','nppv','o2','phyc'), path.copernicusmarine <- "C:/Users/dchagaris/OneDrive - University of Florida/copernicusmarine/copernicusmarine.exe")}
}
