% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.STdriver_gif.R
\name{fn.STdriver_gif}
\alias{fn.STdriver_gif}
\title{Make ST driver gifs}
\usage{
fn.STdriver_gif(
  dir.stdriver = dir.stdriver,
  do.files = NULL,
  datestamps = NULL
)
}
\arguments{
\item{dir.stdriver}{Full directory path to the ST driver folder.}

\item{do.files}{A numeric vector indexing which of the ascii files in dir.stdriver to plot.
The default NULL plots all files in the folder.}

\item{datestamps}{Optional character vector of timesteps for each file.  Setting to NULL
(default) will pull the timestep from the ascii file suffix that is ...YYYYMM.asc.}
}
\value{
Place a gif file in a folder one level up from dri.stdriver
}
\description{
Creates a movie (gif) file from ST driver ascii rasters.
}
\examples{
#example code:
\dontrun{fn.STdriver_gif(dir.stdriver="./ST drivers/30min/wc_vert_int_npp_cefi", do.files=NULL, datestamps=NULL)}
}
