% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fn.STdriver_timeseries.R
\name{fn.STdriver_timeseries}
\alias{fn.STdriver_timeseries}
\title{Get ST driver timeseries average.}
\usage{
fn.STdriver_timeseries(
  dirs.stdriver = dirs.stdriver,
  datestamps = NULL,
  grid.subset = NULL
)
}
\arguments{
\item{dirs.stdriver}{A character vector containing the full directory paths to 1 or more ST
driver variable folders.}

\item{datestamps}{Optional character vector of timesteps for each file.  Setting to NULL
(default) will pull the timestep from the ascii file suffix that is ...YYYYMM.asc.}

\item{grid.subset}{A numeric vector of lenght 4 (xmin,xmax,ymin,ymax) indicating which grid
cells subset and then average over. These are row and column indexes. NULL, the default, averages over the entire spatial grid.  A value of c(1,3,1,3) would get the top left 3x3 corner. Need to switch to a grid mask so cells subset can be any shape/cells.}
}
\value{
Saves a STdriver_timeseries.csv file in the parent directoryr
}
\description{
Creates a movie (gif) file from ST driver ascii rasters.
}
\examples{
#example code:
\dontrun{fn.STdriver_timeseries(dirs.stdriver=list.dirs("./ST drivers/30min")[1:2], grid.subset=NULL, datestamps=NULL)}
}
