% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn_fleetdyn_testval_tags}
\alias{fn_fleetdyn_testval_tags}
\title{Make fleet dynamics parameter taglines for sensitivity runs.}
\usage{
fn_fleetdyn_testval_tags(fleetdyn.base, pct.change = 0.5)
}
\arguments{
\item{fleetdyn.base}{A dataframe with baseline parameters, exported from EwE and read
in the setup file.}

\item{pct.change}{The percent change to use,
applied as pct.change x base and (1+pct.change) x base. Default is 0.5 (+/-50\%).}
}
\value{
A character vector containing the parameter taglines for the command file.
}
\description{
Create parameter lines for the command file to test sensitivity of effective power and effort multiplier parameters.
}
\examples{
# example code:
\dontrun{tags.vul <- fn.vul_testval_tags(predprey=data.frame(pred=1:5,prey=6:10, basevul=2), maxvul=1000, minvul=1.01, is.maxvul.mult=T)}
}
