% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn.envresp_testval_tags}
\alias{fn.envresp_testval_tags}
\title{Make environmental response shape parameter taglines for sensitivity runs.}
\usage{
fn.envresp_testval_tags(env.base, fact = 2)
}
\arguments{
\item{env.base}{A dataframe with baseline mediation parameters, this has to be created manually.}

\item{fact}{Multiplicative factor used to widen and narrow each baseline
response shape when generating the low/high sensitivity taglines. Default 2.}
}
\value{
A character vector containing the parameter taglines for the command file.
}
\description{
Create parameter lines for the command file to test sensitivity to environmental response shapes.
}
\examples{
# example code:
\dontrun{tags.vul <- fn.vul_testval_tags(predprey=data.frame(pred=1:5,prey=6:10, basevul=2), maxvul=1000, minvul=1.01, is.maxvul.mult=T)}
}
