% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Ecospace_GA_calibration.R
\name{fn.GApop}
\alias{fn.GApop}
\title{Generate an initial GA population matrix}
\usage{
fn.GApop(run_config = myconfig, usedist = "unif")
}
\arguments{
\item{run_config}{List containing GA settings; must include \code{popSize}.}

\item{usedist}{Character string; distribution type, either"unif" (uniform between bounds) or
"normln" (log/normal hybrid). Default "unif".}
}
\value{
A numeric matrix of dimension run_config$popSize x n_pars.
}
\description{
Creates the initial population for a genetic algorithm using either
uniform draws within parameter bounds or (optionally) lognormal/normal
draws centered on existing parameter estimates. Uses several global
objects created elsewhere.
}
