% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_Ecospace_GA_calibration.R
\name{fn.parvec2cmd}
\alias{fn.parvec2cmd}
\title{Write command file from parameter vector}
\usage{
fn.parvec2cmd(par_vec = est_par_vec, g = 0, idx = 0, out_dir = run_dir)
}
\arguments{
\item{par_vec}{Parameter vector.}

\item{g}{Generation number, used for file tracking in a genetic algorithm loop. Default 0}

\item{idx}{Index for run identifier.}

\item{out_dir}{Output directory, where the command file will be saved along with model output.#'}
}
\value{
Path to the written command file.
}
\description{
Takes as input a parameter vector and creates taglines and command files.
}
