% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_output.R
\name{fn.read_ecosim_timeseries}
\alias{fn.read_ecosim_timeseries}
\title{Read an ecosim timeseries file.}
\usage{
fn.read_ecosim_timeseries(filename)
}
\arguments{
\item{filename}{The location of the Ecosim timeseries file.}
}
\value{
A list containing an 4 dataframes, one each for observed biomass and catch timeseries and
one each for observed biomass and catch header information.
#examples
#example code:
fn.STdriver_gif(dir.stdriver="./ST drivers/30min/wc_vert_int_npp_cefi", do.files=NULL, datestamps=NULL)
}
\description{
Reads an ecosim timeseries file and parses out the observed (non-forcing) components
}
