% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn.runEwE.parallel}
\alias{fn.runEwE.parallel}
\title{Run EwE CLI app in parallel}
\usage{
fn.runEwE.parallel(
  runlist = runlist,
  obj.fxn = 1,
  cl.export = list("runlist", "obs.ts")
)
}
\arguments{
\item{runlist}{A dataframe that must contain a variable named 'cmd_file', which is the full file path of the command files to run.}

\item{obj.fxn}{A single number indicating the likelihood function to use. 0=none, 1=timeseries only, 2=timeseries and spatial data.}

\item{cl.export}{A list of objects exported to each worker.}
}
\value{
Model output is saved according to command file. If do.obj!=0 then a vector likelihoods is returned and added to the runlist dataframe.
}
\description{
Calls fn.runEwE in parallel framework.
}
\examples{
# example code:
result <- fn.runEwE.parallel(runlist=myrunlist, obj.fxn=1, cl.export=list('myrunlist','obs.ts'))
}
