% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn.runEwE}
\alias{fn.runEwE}
\title{Run EwE CLI app}
\usage{
fn.runEwE(cmdfile, do.obj = 1)
}
\arguments{
\item{cmdfile}{A file path pointing to the command file}

\item{do.obj}{Toggle swith indicating the likelihood function to use. 0=none, 1=timeseries only, 2=timeseries and spatial data}
}
\value{
Model output is saved according to command file. If do.obj!=0 then a vector likelihoods is returned.
}
\description{
Function to run EwE using the CLI
}
\examples{
# example code:
result <- fn.runEwE(cmdfile=C://input//EwEcmd.txt, do.obj=1)
}
