% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R_EwE_console_functions.R
\name{fn.vul_testval_tags}
\alias{fn.vul_testval_tags}
\title{Make vulnerability parameter taglines for sensitivity runs.}
\usage{
fn.vul_testval_tags(predprey, maxvul = 100, minvul = 1.01, is.maxvul.mult = T)
}
\arguments{
\item{predprey}{A dataframe containing 3 numeric columns, "pred", "prey", and "basevul", containing the group numbers of predator prey pairs to evaluate and starting values for the vulnerabilities.  Missing (NA) in the prey column will set the vulnerability by predator column.}

\item{maxvul}{A high vulnerability number to test}

\item{minvul}{A low vulnerability number to test.}

\item{is.maxvul.mult}{Logical indicating whether the maxvul is treated as a multiplier on the basevul (TRUE) or as the absolute value (FALSE).}
}
\value{
A character vector containing the parameter taglines for the command file.
}
\description{
Create parameter taglines for the command file to test sensitivity of vulnerability parameters.
}
\examples{
# example code:
tags.vul <- fn.vul_testval_tags(predprey=data.frame(pred=1:5,prey=6:10, basevul=2), maxvul=1000, minvul=1.01, is.maxvul.mult=T)
}
